Decouple car model
